backend/pyproject.toml
[tool.poetry]
name = "brainrot-plus"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.8.0"
python-dotenv = "^1.0.0"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.26.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
